+++ CHANGELOG.md
@@ -9,8 +9,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Added
 - Detect socket message encodings.
 - All authenticate methods can now pass in username and create options.
+- Support gzip compress/decompress on ApiClient methods.
 
-### Changed
+### Changed
 - Update the code generator to handle POST/DELETE query params.
 - Match listings can now pass through `null` to indicate no filters.
 - ApiClient exceptions now contain HTTP status codes.